Refactor Track
