+++ src/favorites/types/favorites.model.ts
@@ -1,6 +1,6 @@
 import { Artist } from '../../artist/types/artist.model';
 import { Album } from '../../album/types/album.model';
-import { Track } from '../../track/types/track.model';
+import { Track } from '../../track/types/track.type';
 
 export type Favorites = {
   artists: Artist[]; // favorite artists ids
